homework5/requirements.txt
prometheus-client>=0.16.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
openai>=1.0.0
